+++ ChangeLog.md
@@ -5,6 +5,11 @@
 ### [Added]
 - Add a simple example `tensorEye` to create eye tensor.
 - Add an example `simpleNet` to show how to create a network.
+- Add an example `dcgan_struct` to show how to train a network with struct. So far the completion include:
+	- Create the GAN network structure to generate fake MNIST images.
+	- Create a struct to contain training parameters.
+	- Create a network detector to detect MNISt dataset.
+	- Implement a part of the training for-loop to train MNIST detector.
 
 ### [Changed]
 - Update the example `simpleNet` to run network forward.